requirements.txt
beautifulsoup4
python-fasthtml
requests
pydantic